+++ src/kalbach.cpp
@@ -48,7 +48,7 @@ Kalbach::Kalbach(const ACE& ace, std::size_t i)
 
   // Read outgoing energy tables
   for (uint32_t j = 0; j < NE; j++) {
-    uint32_t loc = ace.DLW() + ace.xss<uint32_t>(i + 2 + 2 * NR + NE + j) - 1;
+    uint32_t loc = static_cast<uint32_t>(ace.DLW()) + ace.xss<uint32_t>(i + 2 + 2 * NR + NE + j) - 1;
     try {
       tables_.emplace_back(ace, loc);
     } catch (PNDLException& error) {
@@ -92,7 +92,7 @@ AngleEnergyPacket Kalbach::sample_angle_energy(
     l = incoming_energy_.size() - 2;
     f = 1.;
   } else {
-    l = std::distance(incoming_energy_.begin(), in_E_it) - 1;
+    l = static_cast<std::size_t>(std::distance(incoming_energy_.begin(), in_E_it) - 1);
     f = (E_in - incoming_energy_[l]) /
         (incoming_energy_[l + 1] - incoming_energy_[l]);
   }
@@ -152,7 +152,7 @@ std::optional<double> Kalbach::angle_pdf(double E_in, double mu) const {
     l = incoming_energy_.size() - 2;
     f = 1.;
   } else {
-    l = std::distance(incoming_energy_.begin(), in_E_it) - 1;
+    l = static_cast<std::size_t>(std::distance(incoming_energy_.begin(), in_E_it) - 1);
     f = (E_in - incoming_energy_[l]) /
         (incoming_energy_[l + 1] - incoming_energy_[l]);
   }
@@ -181,7 +181,7 @@ std::optional<double> Kalbach::pdf(double E_in, double mu, double E_out) const {
     l = incoming_energy_.size() - 2;
     f = 1.;
   } else {
-    l = std::distance(incoming_energy_.begin(), in_E_it) - 1;
+    l = static_cast<std::size_t>(std::distance(incoming_energy_.begin(), in_E_it) - 1);
     f = (E_in - incoming_energy_[l]) /
         (incoming_energy_[l + 1] - incoming_energy_[l]);
   }
