+++ src/tabular_energy.cpp
@@ -86,7 +86,7 @@ double TabularEnergy::sample_energy(double E_in,
     l = incoming_energy_.size() - 2;
     f = 1.;
   } else {
-    l = std::distance(incoming_energy_.begin(), in_E_it) - 1;
+    l = static_cast<std::size_t>(std::distance(incoming_energy_.begin(), in_E_it) - 1);
     f = (E_in - incoming_energy_[l]) /
         (incoming_energy_[l + 1] - incoming_energy_[l]);
   }
@@ -129,7 +129,7 @@ std::optional<double> TabularEnergy::pdf(double E_in, double E_out) const {
     l = incoming_energy_.size() - 2;
     f = 1.;
   } else {
-    l = std::distance(incoming_energy_.begin(), in_E_it) - 1;
+    l = static_cast<std::size_t>(std::distance(incoming_energy_.begin(), in_E_it) - 1);
     f = (E_in - incoming_energy_[l]) /
         (incoming_energy_[l + 1] - incoming_energy_[l]);
   }
